front/requirements.txt
websockets==12.0
google-generativeai==0.7.2
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.2
python-dotenv==1.0.0
protobuf==4.25.3